Day1/requirements.txt
Flask==2.3.2
pyJWT==2.8.0